Change Message model.
Drop timestamp - redefined and not really used.
Drop date - unused.
Change created_at type to Float.

+++ test/blueprints.rb
@@ -14,7 +14,7 @@ Message.blueprint do
  facility { rand(15) }
  level { rand(8) }
  host
- created_at { Time.now.to_i }
+ created_at { Time.now.to_f }
  deleted { false }
 end
